Removed the BondComponent interface and renamed BondComponentImpl class to BondComponent
